+++ lib/src/drawer.dart
@@ -7,35 +7,61 @@ class CustomDrawer extends StatelessWidget {
   /// {@macro flutter_widgetz.CustomDrawer}
   const CustomDrawer({
     super.key,
-    required this.itemCount,
-    required this.itemBuilder,
+    required this.delegate,
+    this.color,
     this.header,
   });
 
-  /// The total number of children this delegate can provide.
-  final int itemCount;
+  /// A delegate that provides the children for the [PageView].
+  final SliverChildDelegate delegate;
 
-  /// Called to build children for the sliver.
-  final IndexedWidgetBuilder itemBuilder;
+  /// The background color applied to the [header].
+  final Color? color;
 
   /// A widget to be placed inside the drawer header.
   final Widget? header;
 
+  /// {@macro flutter_widgetz.CustomDrawer}
+  ///
+  /// Count uses [SliverChildListDelegate].
+  CustomDrawer.count({
+    super.key,
+    required List<Widget> children,
+    this.color,
+    this.header,
+  }) : delegate = SliverChildListDelegate(
+          children,
+        );
+
+  /// {@macro flutter_widgetz.CustomDrawer}
+  ///
+  /// Builder uses [SliverChildBuilderDelegate].
+  CustomDrawer.builder({
+    super.key,
+    required int itemCount,
+    required IndexedWidgetBuilder itemBuilder,
+    this.color,
+    this.header,
+  }) : delegate = SliverChildBuilderDelegate(
+          itemBuilder,
+          childCount: itemCount,
+        );
+
   @override
   Widget build(BuildContext context) {
-    return Drawer(
-      child: CustomScrollView(
-        slivers: <Widget>[
-          SliverToBoxAdapter(
-            child: _getHeader(context),
-          ),
-          SliverList(
-            delegate: SliverChildBuilderDelegate(
-              itemBuilder,
-              childCount: itemCount,
+    return SafeArea(
+      top: header == null || false,
+      child: Drawer(
+        child: CustomScrollView(
+          slivers: <Widget>[
+            SliverToBoxAdapter(
+              child: _getHeader(context),
             ),
-          ),
-        ],
+            SliverList(
+              delegate: delegate,
+            ),
+          ],
+        ),
       ),
     );
   }
@@ -47,7 +73,7 @@ class CustomDrawer extends StatelessWidget {
 
     return DrawerHeader(
       decoration: BoxDecoration(
-        color: Theme.of(context).colorScheme.primary,
+        color: color ?? Theme.of(context).colorScheme.primary,
       ),
       child: header,
     );